requirements.txt
fastapi
uvicorn[standard]
pydantic
ebooklib
beautifulsoup4
lxml
python-dotenv
httpx
anthropic
aiofiles
python-multipart
